+++ src/ert/__main__.py
@@ -30,7 +30,7 @@ from ert.config.observation_config_migrations import (
     remove_refcase_and_time_map_dependence_from_obs_config,
 )
 from ert.export_observations import export_observations
-from ert.logging import LOGGING_CONFIG
+from ert.logging.utils import LOGGING_CONFIG
 from ert.mode_definitions import (
     ENIF_MODE,
     ENSEMBLE_EXPERIMENT_MODE,
